+++ ressources/vroomtests.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-vroom --crawl --neovim -u plugin/sniprun.vim --servername=FOO 
-