+++ src/interpreters/Bash_original.rs
@@ -33,7 +33,7 @@ impl Interpreter for Bash_original {
     fn behave_repl_like_default() -> bool {
         true
     }
-    fn has_repl_capability() -> bool{
+    fn has_repl_capability() -> bool {
         true
     }
 
@@ -68,10 +68,10 @@ impl Interpreter for Bash_original {
     fn fetch_code(&mut self) -> Result<(), SniprunError> {
         if !self
             .data
-                .current_bloc
-                .replace(&[' ', '\t', '\n', '\r'][..], "")
-                .is_empty()
-                && self.get_current_level() >= SupportLevel::Bloc
+            .current_bloc
+            .replace(&[' ', '\t', '\n', '\r'][..], "")
+            .is_empty()
+            && self.get_current_level() >= SupportLevel::Bloc
         {
             self.code = self.data.current_bloc.clone();
         } else if !self.data.current_line.replace(" ", "").is_empty()
@@ -108,7 +108,7 @@ impl Interpreter for Bash_original {
             return Ok(String::from_utf8(output.stdout).unwrap());
         } else {
             return Err(SniprunError::RuntimeError(
-                    String::from_utf8(output.stderr).unwrap(),
+                String::from_utf8(output.stderr).unwrap(),
             ));
         }
     }
@@ -195,8 +195,4 @@ mod test_bash_original {
         let string_result = res.unwrap();
         assert_eq!(string_result, "2\n");
     }
-
-
 }
-
-