+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## v0.5.2
+- Ada and Scala interpreters
+- Operator mode, courtesy of @DarwinSenior
+- Julia REPL via jupyter kernel
+- Compatibility fix for nvim 0.4.x (will miss out on features >= sniprun v0.4.9)
+
+
+## v0.5.1 
+- Markdown interpreter
+- Better :SnipInfo
+- SnipInfo \<name> now display the help file in doc/
+
 ## v0.5.0
 - Sniprun becomes a Lua (+Rust) plugin!!
 - Lua backend, config, and checkhealth